# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "EventDto" : {
        "properties" : {
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "category" : {
            "type" : "string"
          },
          "channelId" : {
            "type" : "string"
          },
          "cost" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "maxLength" : 3800,
            "minLength" : 0,
            "type" : "string"
          },
          "displayState" : {
            "type" : "string"
          },
          "guildId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "host" : {
            "type" : "string"
          },
          "hostAvatarUrl" : {
            "type" : "string"
          },
          "hostUsername" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "locationPlaceId" : {
            "type" : "string"
          },
          "messageId" : {
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "notifyOnCreate" : {
            "type" : "boolean"
          },
          "state" : {
            "type" : "string"
          }
        },
        "required" : [ "dateTime", "guildId", "name" ],
        "type" : "object"
      },
...
      "EventUpdateDto" : {
        "properties" : {
          "accepted" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "locationPlaceId" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
...
      "PageMetadata" : {
        "properties" : {
          "number" : {
            "format" : "int64",
            "type" : "integer"
          },
          "size" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PagedModelEventDto" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/EventDto"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        },
        "type" : "object"
      },
...
      "RsvpRequest" : {
        "properties" : {
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/event" : {
      "get" : {
        "description" : "Returns paginated list of non-archived, non-deleted events",
        "operationId" : "getEvents",
        "parameters" : [ {
          "in" : "query",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelEventDto"
                }
              }
            },
            "description" : "Events retrieved successfully"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "List active events",
        "tags" : [ "Events" ]
      },
      "patch" : {
        "description" : "Partially updates event details and syncs with Discord",
        "operationId" : "updateEvent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EventUpdateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Event updated successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Validation error"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Update an existing event",
        "tags" : [ "Events" ]
      },
      "put" : {
        "description" : "Creates an event and its associated Discord channel",
        "operationId" : "createEvent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EventDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Event created successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Validation error"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Create a new event",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}" : {
      "get" : {
        "description" : "Returns full event details including attendee lists",
        "operationId" : "getEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventDetailDto"
                }
              }
            },
            "description" : "Event details retrieved"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Get event details",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}/attendee" : {
      "delete" : {
        "description" : "Removes an attendee from the event. Admins can remove anyone; non-admins can only remove their own +1s.",
        "operationId" : "removeAttendee",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "snowflake",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Attendee removed"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Not authorized to remove this attendee"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Remove an attendee",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}/cancel" : {
      "post" : {
        "description" : "Admin-only: cancels an event and archives it",
        "operationId" : "cancelEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Event cancelled"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Event already completed"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Admin role required"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Cancel an event",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}/private-channel" : {
      "post" : {
        "description" : "Admin-only: creates a private Discord channel for accepted attendees",
        "operationId" : "createPrivateChannel",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Private channel created"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Admin role required"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Create a private channel",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}/recategorize" : {
      "post" : {
        "description" : "Admin-only: triggers a fresh categorization attempt for this event",
        "operationId" : "recategorizeEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Recategorization triggered"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "string"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Admin role required"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "Re-categorize an event",
        "tags" : [ "Events" ]
      }
    },
    "/event/{id}/rsvp" : {
      "post" : {
        "description" : "Record or update your attendance status for an event",
        "operationId" : "rsvpEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RsvpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventDetailDto"
                }
              }
            },
            "description" : "RSVP recorded"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Event not found"
          }
        },
        "summary" : "RSVP to an event",
        "tags" : [ "Events" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}